+++ tests/golden/SecMemCon.sv
@@ -132,7 +132,7 @@ module top_level (input logic [0:0] clk,
   logic [76:0] zi120;
   logic [204:0] zi121;
   logic [204:0] zi122;
-  logic [240:0] main_$l__unused14$856_out;
+  logic [240:0] main_$l__unused14$858_out;
   logic [1:0] zi125;
   logic [9:0] zi126;
   logic [31:0] zi127;
@@ -152,7 +152,7 @@ module top_level (input logic [0:0] clk,
   logic [127:0] zi142;
   logic [204:0] zi143;
   logic [204:0] zi144;
-  logic [240:0] main_$l_main_set_ack_reg$s131$880_out;
+  logic [240:0] main_$l_main_set_ack_reg$s131$882_out;
   logic [9:0] main_addr_reg_outR1;
   logic [9:0] zi148;
   logic [0:0] zi151;
@@ -186,7 +186,7 @@ module top_level (input logic [0:0] clk,
   logic [76:0] zi172;
   logic [204:0] zi173;
   logic [204:0] zi174;
-  logic [240:0] main_$l_main_set_ack_reg$s131$880_outR1;
+  logic [240:0] main_$l_main_set_ack_reg$s131$882_outR1;
   logic [9:0] main_addr_reg_outR2;
   logic [9:0] zi178;
   logic [127:0] main_memtweak_outR1;
@@ -213,9 +213,9 @@ module top_level (input logic [0:0] clk,
   logic [76:0] zi201;
   logic [204:0] zi202;
   logic [204:0] zi203;
-  logic [240:0] main_$l_main_get_addr_reg115$996_out;
+  logic [240:0] main_$l_main_get_addr_reg115$998_out;
   logic [204:0] zi204;
-  logic [240:0] main_$l_main_get_addr_reg115$996_outR1;
+  logic [240:0] main_$l_main_get_addr_reg115$998_outR1;
   logic [240:0] zres;
   assign zi2 = __in0[13:4];
   assign zi3 = zi2;
@@ -341,7 +341,7 @@ module top_level (input logic [0:0] clk,
   assign zi120 = zi119;
   assign zi121 = {zi120, zi114};
   assign zi122 = zi121;
-  main_$L___unused14$856  instR6 (zi122, main_$l__unused14$856_out);
+  main_$L___unused14$858  instR6 (zi122, main_$l__unused14$858_out);
   assign zi125 = __st0[76:75];
   assign zi126 = __st0[74:65];
   assign zi127 = __st0[64:33];
@@ -361,7 +361,7 @@ module top_level (input logic [0:0] clk,
   assign zi142 = main_memtweak_out;
   assign zi143 = {zi140, zi142};
   assign zi144 = zi143;
-  main_$L_Main_set__ack__reg$s131$880  instR9 (zi144, main_$l_main_set_ack_reg$s131$880_out);
+  main_$L_Main_set__ack__reg$s131$882  instR9 (zi144, main_$l_main_set_ack_reg$s131$882_out);
   Main_addr__reg  instR10 (__st0, main_addr_reg_outR1);
   assign zi148 = main_addr_reg_outR1;
   assign zi151 = zi148[9];
@@ -395,7 +395,7 @@ module top_level (input logic [0:0] clk,
   assign zi172 = zi171;
   assign zi173 = {zi172, __st1};
   assign zi174 = zi173;
-  main_$L_Main_set__ack__reg$s131$880  instR19 (zi174, main_$l_main_set_ack_reg$s131$880_outR1);
+  main_$L_Main_set__ack__reg$s131$882  instR19 (zi174, main_$l_main_set_ack_reg$s131$882_outR1);
   Main_addr__reg  instR20 (__st0, main_addr_reg_outR2);
   assign zi178 = main_addr_reg_outR2;
   Main_memTweak  instR21 (zi178, {6'h20{1'h0}}, __st1, main_memtweak_outR1);
@@ -422,10 +422,10 @@ module top_level (input logic [0:0] clk,
   assign zi201 = zi200;
   assign zi202 = {zi201, zi195};
   assign zi203 = zi202;
-  main_$L_Main_get__addr__reg115$996  instR22 (zi203, zi203, main_$l_main_get_addr_reg115$996_out);
+  main_$L_Main_get__addr__reg115$998  instR22 (zi203, zi203, main_$l_main_get_addr_reg115$998_out);
   assign zi204 = {__st0, __st1};
-  main_$L_Main_get__addr__reg115$996  instR23 (zi204, zi204, main_$l_main_get_addr_reg115$996_outR1);
-  assign zres = (__resumption_tag == 3'h0) ? ((zi7 == 1'h1) ? ((zi9 == 1'h0) ? {zi62, 3'h1, zi57} : {zi104, 3'h2, zi99}) : main_$l__unused14$856_out) : ((__resumption_tag == 3'h1) ? main_$l_main_set_ack_reg$s131$880_out : ((__resumption_tag == 3'h2) ? main_$l_main_set_ack_reg$s131$880_outR1 : ((__resumption_tag == 3'h3) ? main_$l_main_get_addr_reg115$996_out : main_$l_main_get_addr_reg115$996_outR1)));
+  main_$L_Main_get__addr__reg115$998  instR23 (zi204, zi204, main_$l_main_get_addr_reg115$998_outR1);
+  assign zres = (__resumption_tag == 3'h0) ? ((zi7 == 1'h1) ? ((zi9 == 1'h0) ? {zi62, 3'h1, zi57} : {zi104, 3'h2, zi99}) : main_$l__unused14$858_out) : ((__resumption_tag == 3'h1) ? main_$l_main_set_ack_reg$s131$882_out : ((__resumption_tag == 3'h2) ? main_$l_main_set_ack_reg$s131$882_outR1 : ((__resumption_tag == 3'h3) ? main_$l_main_get_addr_reg115$998_out : main_$l_main_get_addr_reg115$998_outR1)));
   assign __resumption_tag_next = zres[207:205];
   assign __st0_next = zres[204:128];
   assign __st1_next = zres[127:0];
@@ -440,7 +440,7 @@ module top_level (input logic [0:0] clk,
   end
 endmodule
 
-module main_$L___unused14$856 (input logic [204:0] arg0,
+module main_$L___unused14$858 (input logic [204:0] arg0,
   output logic [240:0] res);
   logic [76:0] zi0;
   logic [76:0] zi1;
@@ -459,7 +459,7 @@ module main_$L___unused14$856 (input logic [204:0] arg0,
   assign res = {zi4, 3'h0, arg0};
 endmodule
 
-module main_$L_Main_set__ack__reg$s131$880 (input logic [204:0] arg0,
+module main_$L_Main_set__ack__reg$s131$882 (input logic [204:0] arg0,
   output logic [240:0] res);
   logic [127:0] zi0;
   logic [1:0] zi1;
@@ -470,7 +470,7 @@ module main_$L_Main_set__ack__reg$s131$880 (input logic [204:0] arg0,
   logic [76:0] zi6;
   logic [204:0] zi7;
   logic [204:0] zi8;
-  logic [240:0] main_$l__unused14$856_out;
+  logic [240:0] main_$l__unused14$858_out;
   assign zi0 = arg0[127:0];
   assign zi1 = arg0[204:203];
   assign zi2 = arg0[202:193];
@@ -480,11 +480,11 @@ module main_$L_Main_set__ack__reg$s131$880 (input logic [204:0] arg0,
   assign zi6 = zi5;
   assign zi7 = {zi6, zi0};
   assign zi8 = zi7;
-  main_$L___unused14$856  inst (zi8, main_$l__unused14$856_out);
-  assign res = main_$l__unused14$856_out;
+  main_$L___unused14$858  inst (zi8, main_$l__unused14$858_out);
+  assign res = main_$l__unused14$858_out;
 endmodule
 
-module main_$L_Main_get__addr__reg115$996 (input logic [204:0] arg0,
+module main_$L_Main_get__addr__reg115$998 (input logic [204:0] arg0,
   input logic [204:0] arg1,
   output logic [240:0] res);
   logic [76:0] zi0;