+++ tests/golden/test3.sv
@@ -4,12 +4,12 @@ module top_level (input logic [0:0] clk,
   output logic [7:0] __out0);
   logic [7:0] __st0;
   logic [7:0] __st0_next;
-  logic [15:0] main_$l_main_tick3$239_out;
-  logic [15:0] main_$l_main_tick3$239_outR1;
+  logic [15:0] main_$l_main_tick3$241_out;
+  logic [15:0] main_$l_main_tick3$241_outR1;
   logic [15:0] zres;
-  main_$L_Main_tick3$239  inst (__st0, main_$l_main_tick3$239_out);
-  main_$L_Main_tick3$239  instR1 (__st0, main_$l_main_tick3$239_outR1);
-  assign zres = (__in0 == 1'h0) ? main_$l_main_tick3$239_out : main_$l_main_tick3$239_outR1;
+  main_$L_Main_tick3$241  inst (__st0, main_$l_main_tick3$241_out);
+  main_$L_Main_tick3$241  instR1 (__st0, main_$l_main_tick3$241_outR1);
+  assign zres = (__in0 == 1'h0) ? main_$l_main_tick3$241_out : main_$l_main_tick3$241_outR1;
   assign __st0_next = zres[7:0];
   assign __out0 = zres[15:8];
   initial __st0 = 8'h0;
@@ -22,7 +22,7 @@ module top_level (input logic [0:0] clk,
   end
 endmodule
 
-module main_$L_Main_tick3$239 (input logic [7:0] arg0,
+module main_$L_Main_tick3$241 (input logic [7:0] arg0,
   output logic [15:0] res);
   assign res = {arg0, arg0};
 endmodule